requirements/requirements_eval_gemini25.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.7.2
anyio==4.11.0
async-lru==2.0.5
asyncio-atexit==1.0.1
attrs==25.4.0
av==16.0.1
backoff==2.2.1
beautifulsoup4==4.14.2
bitsandbytes==0.49.0
black==25.9.0
Brotli==1.1.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
DataProperty==1.1.0
datasets==3.6.0
eva-decord
deepspeed==0.18.1
Deprecated==1.3.1
dill==0.3.8
distlib==0.4.0
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.6
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
ftfy==6.3.1
gdown==5.2.0
gdrive==0.1.5
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.43.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==0.5.3
google-genai==1.47.0
googleapis-common-protos==1.71.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
identify==2.6.15
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflate64==1.0.3
isort==7.0.0
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonlines==4.0.0
kiwisolver==1.4.9
latex2sympy2==1.9.1
lazy_loader==0.4
loguru==0.7.3
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mbstrdecoder==1.1.4
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
multivolumefile==0.2.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nltk==3.9.2
nodeenv==1.9.1
numexpr==2.14.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1; sys_platform == 'linux'
nvidia-cuda-cupti-cu12==12.8.90; sys_platform == 'linux'
nvidia-cuda-nvrtc-cu12==12.8.93; sys_platform == 'linux'
nvidia-cuda-runtime-cu12==12.8.90; sys_platform == 'linux'
nvidia-cudnn-cu12==9.10.2.21; sys_platform == 'linux'
nvidia-cufft-cu12==11.3.3.83; sys_platform == 'linux'
nvidia-cufile-cu12==1.13.1.3; sys_platform == 'linux'
nvidia-curand-cu12==10.3.9.90; sys_platform == 'linux'
nvidia-cusolver-cu12==11.7.3.90; sys_platform == 'linux'
nvidia-cusparse-cu12==12.5.8.93; sys_platform == 'linux'
nvidia-cusparselt-cu12==0.7.1; sys_platform == 'linux'
nvidia-ml-py==13.580.82; sys_platform == 'linux'
nvidia-nccl-cu12==2.27.5; sys_platform == 'linux'
nvidia-nvjitlink-cu12==12.8.93; sys_platform == 'linux'
nvidia-nvshmem-cu12==3.3.20; sys_platform == 'linux'
nvidia-nvtx-cu12==12.8.90; sys_platform == 'linux'
oauthlib==3.3.1
openai==2.6.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-instrumentation-urllib==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pathvalidate==3.3.1
peft==0.17.1
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
pre_commit==4.3.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==5.9.8
py-cpuinfo==9.0.0
py7zr==1.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybcj==1.0.6
pybind11==3.0.1
pycocoevalcap==1.2
pycocotools==2.0.10
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pynrrd==1.1.3
pyparsing==3.2.5
pyppmd==1.2.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
pyzstd==0.18.0
qwen-vl-utils==0.0.14
rarfile==4.2
regex==2025.10.23
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sacrebleu==2.5.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
SecretStorage==3.3.3
sentence-transformers==5.1.2
sentencepiece==0.2.1
sentry-sdk==2.43.0
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sqlitedict==2.1.0
sympy==1.14.0
synapseclient==4.10.0
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tenacity==9.1.2
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tiktoken==0.12.0
timm==1.0.21
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.57.1
transformers-stream-generator==0.0.5
triton==3.5.0; sys_platform == 'linux'
typepy==1.3.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==1.26.20
versioneer==0.29
virtualenv==20.35.4
wandb==0.22.3
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
yt-dlp==2025.10.22
zipp==3.23.0
zss==1.2.0
zstandard==0.25.0
